+++ Emulator/Agnus/Copper/Copper.cpp
@@ -134,7 +134,7 @@ Copper::findVerticalMatch(i16 vStrt, i16 vComp, i16 vMask, i16 &result) const
     i16 vStop = agnus.frame.numLines();
 
     // Iterate through all vertical positions
-    for (int v = vStrt; v < vStop; v++) {
+    for (isize v = vStrt; v < vStop; v++) {
 
         // Check if the comparator triggers at this position
         if ((v & vMask) >= (vComp & vMask)) {
@@ -151,7 +151,7 @@ Copper::findHorizontalMatch(i16 hStrt, i16 hComp, i16 hMask, i16 &result) const
     i16 hStop = HPOS_CNT;
 
     // Iterate through all horizontal positions
-    for (int h = hStrt; h < hStop; h++) {
+    for (isize h = hStrt; h < hStop; h++) {
 
         // Check if the comparator triggers at this position
         if ((h & hMask) >= (hComp & hMask)) {
